[MODIFY]:리덕스 제거
