+++ package.json
@@ -17,7 +17,6 @@
         "@emotion/css": "^11.11.0",
         "@emotion/react": "^11.11.0",
         "@emotion/styled": "^11.11.0",
-        "@reduxjs/toolkit": "^1.9.5",
         "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
@@ -29,14 +28,12 @@
         "mobx": "^6.7.0",
         "mobx-react": "^7.6.0",
         "object-to-formdata": "^4.5.1",
-        "query-string": "^8.1.0",
         "react": "^18.2.0",
         "react-calendar": "^4.3.0",
         "react-daum-postcode": "^3.1.1",
         "react-dom": "^18.2.0",
         "react-hook-form": "^7.44.3",
         "react-icons": "^4.9.0",
-        "react-redux": "^8.0.7",
         "react-router-dom": "^6.11.2",
         "react-scripts": "5.0.1",
         "react-toastify": "^9.1.3",
