--- src/api/@utils/statusCode.ts
@@ -1,15 +1,0 @@
-const OK = 200;
-const BAD_REQUEST = 400;
-const UNAUTHORIZED = 401;
-const FORBIDDEN = 403;
-const NOT_FOUND = 404;
-const INTERNAL_SERVER_ERROR = 500;
-
-export default {
-  OK,
-  BAD_REQUEST,
-  UNAUTHORIZED,
-  FORBIDDEN,
-  NOT_FOUND,
-  INTERNAL_SERVER_ERROR,
-};
