--- src/api/@utils/statusMessage.ts
@@ -1,27 +1,0 @@
-const ACCESS_TOKEN_EXPIRED = "accessToken Expired";
-const ACCESS_TOKEN_SIGNATURE_EXCEPTION = "accessToken SignatureException";
-const ACCESS_TOKEN_MALFORMED = "accessToken Malformed";
-const REFRESH_TOKEN_EXPIRED = "refreshToken Expired";
-const REFRESH_TOKEN_SIGNATURE_EXCEPTION = "refreshToken SignatureException";
-const REFRESH_TOKEN_MALFORMED = "refreshToken Malformed";
-const INTERNAL_SERVER_ERROR_MESSAGE = "Internal Server Error";
-
-const NOT_FOUND_USER_VIA_MONSTER_CODE = "Not Found User";
-const USER_DELETED = "User Droped";
-const USER_NAME_UPDATE_SUCCESS = "User Name Updated Completed";
-
-const MONSTER_NAME_UPDATE_SUCCESS = "Change Monster Name";
-
-export default {
-  ACCESS_TOKEN_EXPIRED,
-  ACCESS_TOKEN_MALFORMED,
-  ACCESS_TOKEN_SIGNATURE_EXCEPTION,
-  REFRESH_TOKEN_EXPIRED,
-  REFRESH_TOKEN_MALFORMED,
-  REFRESH_TOKEN_SIGNATURE_EXCEPTION,
-  INTERNAL_SERVER_ERROR_MESSAGE,
-  NOT_FOUND_USER_VIA_MONSTER_CODE,
-  USER_DELETED,
-  USER_NAME_UPDATE_SUCCESS,
-  MONSTER_NAME_UPDATE_SUCCESS,
-};
