+++ src/pages/FeedDetail/BottomSection.tsx
@@ -1,23 +1,17 @@
 import React from 'react';
-
-import { useSelector } from 'react-redux';
-import { RootState } from 'redux/store';
-
 import Box from 'components/atoms/Box/default-box';
 import Review from 'components/organisms/Review';
 
 import feedDetail from 'db/feedDetail.json';
+import review from 'db/review.json';
 
 const BottomSection = () => {
-    const { reviewShow } = useSelector((state: RootState) => state.toggle);
-    const { reviewList } = useSelector((state: RootState) => state.user);
-
     return (
         <>
-            {reviewShow && (
+            {true && (
                 <Box borderWidth={1} padding="5px" marginTop={10} width={100} height={100}>
                     <Review variant="me" author={feedDetail.author!} id={feedDetail.id!} />
-                    {reviewList.map((review, index) => (
+                    {review.data.map((review, index) => (
                         <Review key={index} variant="people" {...review} />
                     ))}
                 </Box>
